analyze: select parts of the graph to hide
Add options to systemd-analyze plot allowing to hide statrting parts
the graph before the main PID#1 systemd.

+++ src/analyze/analyze-plot.c
@@ -130,6 +130,12 @@ static int compare_unit_start(const UnitTimes *a, const UnitTimes *b) {
 }
 
 static void svg_graph_box(double height, double begin, double end) {
+        /* show timestamps relative to the beginning of the plot
+         * keep the firmware and loader timestamps at the negative side */
+        if (begin > 0) {
+                end = end - begin;
+                begin = 0;
+        }
         /* outside box, fill */
         svg("<rect class=\"box\" x=\"%.03f\" y=\"0\" width=\"%.03f\" height=\"%.03f\" />\n",
             SCALE_X * (begin),
@@ -181,7 +187,7 @@ static void plot_tooltip(const UnitTimes *ut) {
                 }
 }
 
-static int plot_unit_times(UnitTimes *u, double width, int y) {
+static int plot_unit_times(UnitTimes *u, double width, double graph_start_time, int y) {
         bool b;
 
         if (!u->name)
@@ -196,7 +202,7 @@ static int plot_unit_times(UnitTimes *u, double width, int y) {
         svg_bar("deactivating", u->deactivating, u->deactivated, y);
 
         /* place the text on the left if we have passed the half of the svg width */
-        b = u->activating * SCALE_X < width / 2;
+        b = (u->activating - graph_start_time) * SCALE_X < width / 2;
         if (u->time)
                 svg_text(b, u->activating, y, "%s (%s)",
                          u->name, FORMAT_TIMESPAN(u->time, USEC_PER_MSEC));
@@ -228,20 +234,31 @@ static int produce_plot_as_svg(
         UnitTimes *u;
         double width;
 
-        width = SCALE_X * (boot->firmware_time + boot->finish_time);
+        double graph_start_time = -(double) boot->firmware_time;
+        if ((arg_plot_hide_mode == HIDE_SOFT_REBOOT && boot->soft_reboots_count > 0) ||
+            arg_plot_hide_mode == HIDE_INITRD)
+                graph_start_time = (double) boot->userspace_time;
+        else if (arg_plot_hide_mode == HIDE_KERNEL)
+                graph_start_time = (double) boot->kernel_done_time;
+        else if (arg_plot_hide_mode == HIDE_LOADER)
+                graph_start_time = 0;
+        else if (arg_plot_hide_mode == HIDE_FIRMWARE)
+                graph_start_time = -(double) boot->loader_time;
+
+        width = SCALE_X * (boot->finish_time - graph_start_time);
         if (width < 800.0)
                 width = 800.0;
 
-        if (boot->firmware_time > boot->loader_time)
+        if (arg_plot_hide_mode < HIDE_FIRMWARE && boot->firmware_time > boot->loader_time)
                 m++;
-        if (timestamp_is_set(boot->loader_time)) {
+        if (arg_plot_hide_mode < HIDE_LOADER && timestamp_is_set(boot->loader_time)) {
                 m++;
                 if (width < 1000.0)
                         width = 1000.0;
         }
-        if (timestamp_is_set(boot->initrd_time))
+        if (arg_plot_hide_mode < HIDE_INITRD && timestamp_is_set(boot->initrd_time))
                 m++;
-        if (timestamp_is_set(boot->kernel_done_time))
+        if (arg_plot_hide_mode < HIDE_KERNEL && timestamp_is_set(boot->kernel_done_time))
                 m++;
 
         for (u = times; u->has_data; u++) {
@@ -320,28 +337,32 @@ static int produce_plot_as_svg(
                     strempty(host->architecture),
                     strempty(host->virtualization));
 
-        svg("<g transform=\"translate(%.3f,100)\">\n", 20.0 + (SCALE_X * boot->firmware_time));
-        if (boot->soft_reboots_count > 0)
+        svg("<g transform=\"translate(%.3f,100)\">\n",
+            20.0 - (SCALE_X * (graph_start_time < 0 ? graph_start_time : 0)));
+        if (arg_plot_hide_mode != HIDE_SOFT_REBOOT && boot->soft_reboots_count > 0)
                 svg_graph_box(m, 0, boot->finish_time);
         else
-                svg_graph_box(m, -(double) boot->firmware_time, boot->finish_time);
+                svg_graph_box(m, graph_start_time, boot->finish_time);
 
-        if (timestamp_is_set(boot->firmware_time)) {
+        if (arg_plot_hide_mode < HIDE_FIRMWARE && timestamp_is_set(boot->firmware_time)) {
                 svg_bar("firmware", -(double) boot->firmware_time, -(double) boot->loader_time, y);
                 svg_text(true, -(double) boot->firmware_time, y, "firmware");
                 y++;
         }
-        if (timestamp_is_set(boot->loader_time)) {
+        if (arg_plot_hide_mode < HIDE_LOADER && timestamp_is_set(boot->loader_time)) {
                 svg_bar("loader", -(double) boot->loader_time, 0, y);
                 svg_text(true, -(double) boot->loader_time, y, "loader");
                 y++;
         }
-        if (timestamp_is_set(boot->kernel_done_time)) {
+        if (arg_plot_hide_mode < HIDE_KERNEL && timestamp_is_set(boot->kernel_done_time)) {
                 svg_bar("kernel", 0, boot->kernel_done_time, y);
                 svg_text(true, 0, y, "kernel");
                 y++;
         }
-        if (timestamp_is_set(boot->initrd_time)) {
+
+        svg("<g transform=\"translate(%.3f,0)\">\n", -(SCALE_X * (graph_start_time > 0 ? graph_start_time : 0)));
+
+        if (arg_plot_hide_mode < HIDE_INITRD && timestamp_is_set(boot->initrd_time)) {
                 svg_bar("initrd", boot->initrd_time, boot->userspace_time, y);
                 if (boot->initrd_security_start_time < boot->initrd_security_finish_time)
                         svg_bar("security", boot->initrd_security_start_time, boot->initrd_security_finish_time, y);
@@ -352,7 +373,7 @@ static int produce_plot_as_svg(
                 svg_text(true, boot->initrd_time, y, "initrd");
                 y++;
         }
-        if (boot->soft_reboots_count > 0) {
+        if (arg_plot_hide_mode != HIDE_SOFT_REBOOT && boot->soft_reboots_count > 0) {
                 svg_bar("soft-reboot", 0, boot->userspace_time, y);
                 svg_text(true, 0, y, "soft-reboot");
                 y++;
@@ -362,7 +383,7 @@ static int produce_plot_as_svg(
                 if (u->activating >= boot->userspace_time)
                         break;
 
-                y += plot_unit_times(u, width, y);
+                y += plot_unit_times(u, width, graph_start_time, y);
         }
 
         svg_bar("active", boot->userspace_time, boot->finish_time, y);
@@ -376,9 +397,9 @@ static int produce_plot_as_svg(
         y++;
 
         for (; u->has_data; u++)
-                y += plot_unit_times(u, width, y);
+                y += plot_unit_times(u, width, graph_start_time, y);
 
-        svg("</g>\n");
+        svg("</g></g>\n");
 
         /* Legend */
         svg("<g transform=\"translate(20,100)\">\n");
